Project store, router, and views moved into modules/project
