+++ src/modules/project/views/show/components/Contributors.vue
