+++ src/modules/project/views/show/components/layout.vue
