+++ src/modules/project/views/show/index.vue
